[Update] BallSprite: Collision Moved
The collision check that was here was moved to the ObjectWall class, also added support for the ObjectWallFinale class as to whether or not it collided. Whether or not it collided with a wall and what wall is returned into this sub now, where it takes the action to corresponding wall.